+++ CHANGES.rst
@@ -30,6 +30,7 @@ in progress
 - [grafana] Use specific Grafana folder for stuffing instant dashboards into
 - [grafana] Use topology information for deriving unique dashboard id from
 - [grafana] Add mechanisms for taming the dashboard refresh interval
+- [core] Improve component loading robustness
 
 
 .. _kotori-0.20.1: